refactor: isolate semiring compile path and mark non-mainline (design_v3 Stage 2)
- Added #[deprecated] to SemiringOp, SemiringBackend, SemiringOps per
  design_v3/30-algebra-and-tropical.md "Recommended Fate Of SemiringOp".
- Moved the semiring compile path in tenferro/src/compiler.rs into a
  dedicated submodule; reduced visibility to pub(crate).
- Narrowed #[allow(deprecated)] to internal call sites that still use
  these items for legacy test / compatibility paths.
- SemiringOp pipeline remains functional until Stage 6 removal.
