+++ tenferro-tensor/src/cubecl/mod.rs
@@ -77,7 +77,13 @@ use cubecl_cuda::CudaRuntime;
 use num_complex::{Complex32, Complex64};
 use tenferro_algebra::Semiring;
 
-use crate::backend::{SemiringBackend, TensorBackend};
+use crate::backend::TensorBackend;
+// `SemiringBackend` is the legacy algebra-generic backend trait, deprecated in
+// design_v3 Stage 2 and scheduled for removal in Stage 6. The impl below is
+// retained so cubecl builds continue to compile; suppress the import-level
+// warning at the narrow scope of the import.
+#[allow(deprecated)]
+use crate::backend::SemiringBackend;
 use crate::config::{
     CompareDir, DotGeneralConfig, GatherConfig, PadConfig, ScatterConfig, SliceConfig,
 };
@@ -2518,6 +2524,7 @@ impl TensorBackend for CubeclBackend {
     }
 }
 
+#[allow(deprecated)]
 impl<Alg: Semiring> SemiringBackend<Alg> for CubeclBackend {
     fn batched_gemm(
         &mut self,
