+++ tenferro-tensor/src/cuda/mod.rs
@@ -1,6 +1,12 @@
 use tenferro_algebra::Semiring;
 
-use crate::backend::{SemiringBackend, TensorBackend};
+use crate::backend::TensorBackend;
+// `SemiringBackend` is the legacy algebra-generic backend trait, deprecated in
+// design_v3 Stage 2 and scheduled for removal in Stage 6. The stubs below are
+// retained so CUDA builds continue to compile; suppress the import-level
+// warning at the narrow scope of the import.
+#[allow(deprecated)]
+use crate::backend::SemiringBackend;
 use crate::config::{
     CompareDir, DotGeneralConfig, GatherConfig, PadConfig, ScatterConfig, SliceConfig,
 };
@@ -229,6 +235,7 @@ macro_rules! impl_stub_backend {
             }
         }
 
+        #[allow(deprecated)]
         impl<Alg: Semiring> SemiringBackend<Alg> for $ty {
             fn batched_gemm(
                 &mut self,
