+++ tenferro-tensor/tests/runtime_error_tests.rs
@@ -1,3 +1,8 @@
+// Test file exercises the deprecated legacy `SemiringBackend` trait; suppress
+// the deprecation warnings at module scope rather than annotating each
+// assertion.
+#![allow(deprecated)]
+
 use std::panic::{catch_unwind, AssertUnwindSafe};
 
 use tenferro_algebra::Standard;
